+++ src/init/mod.rs
@@ -5,4 +5,12 @@ pub mod application_feature;
 pub mod rosella_features;
 mod utils;
 
+pub use rosella_features::register_rosella_headless;
+pub use rosella_features::register_rosella_debug;
+
 pub use initialization_registry::InitializationRegistry;
+
+pub use application_feature::ApplicationInstanceFeature;
+pub use application_feature::ApplicationDeviceFeature;
+pub use application_feature::ApplicationDeviceFeatureGenerator;
+pub use application_feature::FeatureAccess;
